+++ R/pkg_tutorials.R
@@ -25,6 +25,22 @@ run_tute <- function(tute) {
   } else {
     try(learnr::run_tutorial(tute, "manynet"), silent = TRUE)
     try(learnr::run_tutorial(tute, "migraph"), silent = TRUE)
+    cat("Didn't find a direct match, so looking for close matches...")
+    t1 <- dplyr::as_tibble(learnr::available_tutorials(package = "manynet"),
+                           silent = TRUE) %>% dplyr::select(1:3)
+    t2 <- dplyr::as_tibble(learnr::available_tutorials(package = "migraph"),
+                           silent = TRUE) %>% dplyr::select(1:3)
+    avails <- rbind(t1, t2)
+    inftit <- grepl(tute, avails$title, ignore.case = TRUE)
+    if(!any(inftit) | sum(inftit)>1)
+      inftit <- which.min(adist(tute, avails$title, ignore.case = TRUE))
+    if(any(inftit) & sum(inftit)==1){
+      cat(" and found one!")
+      try(learnr::run_tutorial(avails$name[inftit], avails$package[inftit]), silent = TRUE)
+    } else{
+      cat(" and couldn't find which one you meant. Please specify one of these titles:\n")
+      print(avails)
+    }
   }
 }